FTR: move basic services under common folder (#66563)
* [CODEOWNERS] set kibana-qa to own test/functional/services

* [functional/services] move low-level services to common folder

* update codeowners

* update imports

# Conflicts:
#	.github/CODEOWNERS
