+++ test/functional/services/common/test_subjects.ts
@@ -20,8 +20,8 @@
 import testSubjSelector from '@kbn/test-subj-selector';
 import { map as mapAsync } from 'bluebird';
 import { ProvidedType } from '@kbn/test/types/ftr';
-import { WebElementWrapper } from './lib/web_element_wrapper';
-import { FtrProviderContext } from '../ftr_provider_context';
+import { WebElementWrapper } from '../lib/web_element_wrapper';
+import { FtrProviderContext } from '../../ftr_provider_context';
 
 interface ExistsOptions {
   timeout?: number;
